+++ src/App/Resources/skeleton/doctrine/Entity.tpl.php
@@ -1,0 +1,36 @@
+<?= "<?php\n" ?>
+
+/*
+ * This file is part of API as a Service.
+ *
+ * Copyright (c) 2019 Christian Siewert <christian@sieware.international>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace <?= $namespace ?>;
+
+<?php if ($api_resource): ?>use ApiPlatform\Core\Annotation\ApiResource;
+<?php endif ?>
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+<?php if ($api_resource): ?> * @ApiResource(routePrefix="/api")
+<?php endif ?>
+ * @ORM\Entity(repositoryClass="<?= $repository_full_class_name ?>")
+ */
+class <?= $class_name."\n" ?>
+{
+    /**
+     * @ORM\Id()
+     * @ORM\GeneratedValue()
+     * @ORM\Column(type="integer")
+     */
+    private $id;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+}