+++ src/App/Resources/skeleton/doctrine/TreeEntity.tpl.php
@@ -1,5 +1,14 @@
 <?= "<?php\n" ?>
 
+/*
+ * This file is part of API as a Service.
+ *
+ * Copyright (c) 2019 Christian Siewert <christian@sieware.international>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace <?= $namespace ?>;
 
 <?php if ($api_resource): ?>use ApiPlatform\Core\Annotation\ApiResource;